bot/requirements.txt
pinhook==1.7.1
flair==0.4.0
pyyaml==3.13
imgurpython==1.1.7
